move mermaid and graphviz into seperate compilation units
