+++ CMakeLists.txt
@@ -24,7 +24,8 @@ add_executable(gitmem
   src/interpreter.cc
   src/debugger.cc
   src/model_checker.cc
-  src/graph.cc
+  src/graphviz.cc
+  src/mermaid.cc
 )
 
 add_executable(gitmem_trieste
